Fix sign comparison warning

+++ src/utils/mtev_watchdog.c
@@ -819,7 +819,7 @@ mtev_setup_crash_signals(void (*action)(int, siginfo_t *, void *)) {
     altstack_size = default_altstack_size = atoi(envcp);
   }
   if(default_altstack_size > 0 && alt_stack_ptr == NULL) {
-    altstack_size = MAX(MINSIGSTKSZ, default_altstack_size);
+    altstack_size = MAX(MINSIGSTKSZ, (long) default_altstack_size);
     if((altstack.ss_sp = malloc(altstack_size)) == NULL)
       altstack_size = 0;
     else {
